refactor(lib): merge lib/versions.ts into lib/git.ts
